refactor(benchmarks): refactor to support AOT bootstrap in G3 (#12075)
