+++ modules/benchmarks/src/largetable/baseline/index.ts
@@ -2,24 +2,28 @@ import {bindAction, profile} from '../../util';
 import {buildTable, emptyTable} from '../util';
 import {TableComponent} from './table';
 
-export function main() {
-  var table: TableComponent;
+var table: TableComponent;
 
-  function destroyDom() { table.data = emptyTable; }
-
-  function createDom() { table.data = buildTable(); }
-
-  function noop() {}
-
-  function init() {
-    table = new TableComponent(document.querySelector('largetable'));
-
-    bindAction('#destroyDom', destroyDom);
-    bindAction('#createDom', createDom);
-
-    bindAction('#updateDomProfile', profile(createDom, noop, 'update'));
-    bindAction('#createDomProfile', profile(createDom, destroyDom, 'create'));
-  }
-
-  init();
+function destroyDom() {
+  table.data = emptyTable;
 }
+
+function createDom() {
+  table.data = buildTable();
+}
+
+function noop() {}
+
+function init() {
+  const rootEl = document.querySelector('largetable');
+  rootEl.textContent = '';
+  table = new TableComponent(rootEl);
+
+  bindAction('#destroyDom', destroyDom);
+  bindAction('#createDom', createDom);
+
+  bindAction('#updateDomProfile', profile(createDom, noop, 'update'));
+  bindAction('#createDomProfile', profile(createDom, destroyDom, 'create'));
+}
+
+init();
