+++ modules/benchmarks/src/tree/baseline/index.html
@@ -19,9 +19,13 @@
   </p>
 
   <div>
-    <tree id="root"></tree>
+    <tree id="root">Loading...</tree>
   </div>
 
-  <script src="../../bootstrap_plain.js"></script>
+  <script>
+    var mainUrl = window.location.search.split(/[?&]main=([^&]+)/)[1]
+      || '../../bootstrap_plain.js';
+    document.write('<script src="' + mainUrl + '">\u003c/script>');
+  </script>
 </body>
 </html>
