+++ modules/benchmarks/src/tree/baseline/index.ts
@@ -2,24 +2,28 @@ import {bindAction, profile} from '../../util';
 import {buildTree, emptyTree} from '../util';
 import {TreeComponent} from './tree';
 
-export function main() {
-  var tree: TreeComponent;
+var tree: TreeComponent;
 
-  function destroyDom() { tree.data = emptyTree; }
-
-  function createDom() { tree.data = buildTree(); }
-
-  function noop() {}
-
-  function init() {
-    tree = new TreeComponent(document.querySelector('tree'));
-
-    bindAction('#destroyDom', destroyDom);
-    bindAction('#createDom', createDom);
-
-    bindAction('#updateDomProfile', profile(createDom, noop, 'update'));
-    bindAction('#createDomProfile', profile(createDom, destroyDom, 'create'));
-  }
-
-  init();
+function destroyDom() {
+  tree.data = emptyTree;
 }
+
+function createDom() {
+  tree.data = buildTree();
+}
+
+function noop() {}
+
+function init() {
+  const rootEl = document.querySelector('tree');
+  rootEl.textContent = '';
+  tree = new TreeComponent(rootEl);
+
+  bindAction('#destroyDom', destroyDom);
+  bindAction('#createDom', createDom);
+
+  bindAction('#updateDomProfile', profile(createDom, noop, 'update'));
+  bindAction('#createDomProfile', profile(createDom, destroyDom, 'create'));
+}
+
+init();
